Creating helpers class for _seq_str and _switch_device
