+++ pulser-core/pulser/sequence/helpers/_seq_str.py
